requirements.txt
paddlepaddle==2.6.1
paddleocr==2.8.1
surya-ocr
